Fix font weight issue in editor
The editor doesn't behave properly when using variables for font weight. Notably, when you go to change the font weight in global styles, the font weight defaults to a very thin font. This is likely due to it not reading the variable or expecting a number value.

I've switched this to use the number value for each font weight.